PR comments
* Flexible sync only for geospatial
* Convenience function for tear down
* Add error case for altitude
* Refactor altitude test to box, since circle is erroneous from baas
