+++ integration-tests/tests/src/tests.ts
@@ -28,6 +28,7 @@ import "./tests/sync/client-reset";
 import "./tests/sync/dictionary";
 import "./tests/sync/encryption";
 import "./tests/sync/flexible";
+import "./tests/sync/geospatial";
 import "./tests/sync/mixed";
 import "./tests/sync/mongo-db-client";
 import "./tests/sync/open-behavior";
