+++ integration-tests/tests/src/tests/queries.ts
@@ -15,23 +15,10 @@
 // limitations under the License.
 //
 ////////////////////////////////////////////////////////////////////////////
-import Realm, {
-  BSON,
-  GeoBox,
-  GeoCircle,
-  GeoPolygon,
-  CanonicalGeoPoint,
-  GeoPosition,
-  CanonicalGeoPolygon,
-  kmToRadians,
-  miToRadians,
-  WaitForSync,
-} from "realm";
-import { expect, use } from "chai";
-import { authenticateUserBefore, importAppBefore, openRealmBeforeEach } from "../hooks";
+import Realm, { BSON, WaitForSync } from "realm";
+import { expect } from "chai";
+import { openRealmBeforeEach } from "../hooks";
 import { ContactSchema, IContact } from "../schemas/contact";
-import { buildAppConfig } from "../utils/build-app-config";
-import { OpenRealmConfiguration } from "../utils/open-realm";
 
 interface INullableTypesObject {
   boolCol?: boolean;
@@ -150,46 +137,6 @@ class Story extends Realm.Object implements IStory {
   };
 }
 
-class MyGeoPoint implements CanonicalGeoPoint {
-  coordinates: GeoPosition;
-  type = "Point" as const;
-
-  constructor(long: number, lat: number) {
-    this.coordinates = [long, lat];
-  }
-
-  static schema: Realm.ObjectSchema = {
-    name: "MyGeoPoint",
-    embedded: true,
-    properties: {
-      type: "string",
-      coordinates: "double[]",
-    },
-  };
-}
-
-interface IPointOfInterest {
-  _id: number;
-  location: MyGeoPoint;
-  locations?: MyGeoPoint[];
-}
-
-class PointOfInterest extends Realm.Object implements IPointOfInterest {
-  _id = 0;
-  location = new MyGeoPoint(0, 0);
-  locations = [new MyGeoPoint(0, 0)];
-
-  static schema: Realm.ObjectSchema = {
-    name: "PointOfInterest",
-    properties: {
-      _id: "int",
-      location: "MyGeoPoint",
-      locations: "MyGeoPoint[]",
-    },
-    primaryKey: "_id",
-  };
-}
-
 type QueryLengthPair = [ExpectedLength: number, Query: string, ...QueryArgs: Array<any>];
 type QueryExceptionPair = [ExpectedException: string, Query: string, ...QueryArgs: Array<any>];
 type QueryResultsPair = [ExpectedResults: any[], Query: string, ...QueryArgs: Array<any>];
@@ -254,15 +201,11 @@ const expectQueryResultValues = async (
   objectSchema: Realm.ObjectClass,
   propertyToCompare: string,
   queryResultPairs: QueryResultsPair[],
-  useFlexibleSync = false,
 ) => {
   return Promise.all(
     queryResultPairs.map(async ([expectedResults, queryString, ...queryArgs]) => {
       let results = realm.objects(objectSchema);
       results = results.filtered(queryString, ...queryArgs);
-      if (useFlexibleSync) {
-        await results.subscribe({ behavior: WaitForSync.Always });
-      }
       expect(results.length).to.equal(expectedResults.length);
       expect(results.map((el) => (el as any)[propertyToCompare])).to.deep.equal(
         expectedResults,
@@ -270,11 +213,6 @@ const expectQueryResultValues = async (
       Expected results not returned from query: ${queryString} ${JSON.stringify(queryArgs)},
     `,
       );
-      if (useFlexibleSync) {
-        results.unsubscribe();
-        await realm.syncSession?.downloadAllServerChanges();
-        expect(results.length).to.equal(0);
-      }
     }),
   );
 };
@@ -284,14 +222,10 @@ const expectQueryResultValuesByName = (
   objectSchemaName: string,
   propertyToCompare: string,
   queryResultPairs: QueryResultsPair[],
-  useFlexibleSync = false,
 ) => {
   return Promise.all(
     queryResultPairs.map(async ([expectedResults, queryString, ...queryArgs]) => {
       let results = realm.objects(objectSchemaName);
-      if (useFlexibleSync) {
-        await results.subscribe({ behavior: WaitForSync.Always });
-      }
       results = results.filtered(queryString, ...queryArgs);
       expect(results.length).to.equal(expectedResults.length);
       expect(results.map((el) => (el as any)[propertyToCompare])).to.deep.equal(
@@ -300,926 +234,11 @@ const expectQueryResultValuesByName = (
       Expected results not returned from query: ${queryString} ${JSON.stringify(queryArgs)},
     `,
       );
-      if (useFlexibleSync) {
-        results.unsubscribe();
-        await realm.syncSession?.downloadAllServerChanges();
-        expect(results.length).to.equal(0);
-      }
     }),
   );
 };
 
 describe("Queries", () => {
-  [false, true].forEach((useFlexibleSync) => {
-    describe.skipIf(
-      environment.missingServer && useFlexibleSync,
-      `GeoSpatial ${useFlexibleSync ? "(with QBS)" : ""}`,
-      () => {
-        useFlexibleSync && importAppBefore(buildAppConfig("with-flx").anonAuth().flexibleSync());
-        useFlexibleSync && authenticateUserBefore();
-        const realmConfig = (useFlexibleSync
-          ? { schema: [PointOfInterest, MyGeoPoint.schema], sync: { flexible: true } }
-          : { schema: [PointOfInterest, MyGeoPoint.schema] }) as unknown as OpenRealmConfiguration;
-        openRealmBeforeEach(realmConfig);
-        const zero: IPointOfInterest = {
-          _id: 1,
-          location: new MyGeoPoint(0, 0),
-        };
-
-        const poiA: IPointOfInterest = {
-          _id: 2,
-          location: new MyGeoPoint(50.5, 50.5),
-        };
-
-        const poiB: IPointOfInterest = {
-          _id: 3,
-          location: new MyGeoPoint(40, -40),
-        };
-
-        const poiC: IPointOfInterest = {
-          _id: 4,
-          location: new MyGeoPoint(-30, -30.5),
-        };
-
-        const poiD: IPointOfInterest = {
-          _id: 5,
-          location: new MyGeoPoint(-25, 25),
-        };
-
-        const northPole: IPointOfInterest = {
-          _id: 6,
-          location: new MyGeoPoint(0.01, 89.9),
-        };
-
-        const invalid: IPointOfInterest = {
-          _id: 7,
-          location: new MyGeoPoint(2129.01, 89.9),
-        };
-
-        async function geoTest(
-          realm: Realm,
-          geo: GeoCircle | GeoBox | GeoPolygon,
-          pois: IPointOfInterest[],
-          useFlexibleSync = false,
-        ) {
-          const poiIds = pois.map((p) => p._id);
-          return expectQueryResultValues(
-            realm,
-            PointOfInterest,
-            "_id",
-            [[poiIds, "location geoWithin $0 SORT(_id ASC)", geo]],
-            useFlexibleSync,
-          );
-        }
-
-        function geoException(realm: Realm, geo: GeoCircle | GeoBox | GeoPolygon, exception: string) {
-          expectQueryException(realm, PointOfInterest, [[exception, "location geoWithin $0 SORT(_id ASC)", geo]]);
-        }
-
-        beforeEach(async function (this: RealmContext | UserContext) {
-          const objects = this.realm.objects(PointOfInterest);
-          if (useFlexibleSync) {
-            await objects.subscribe();
-          }
-          if (objects.length === 0) {
-            this.realm.write(() => {
-              this.realm.create(PointOfInterest, zero);
-              this.realm.create(PointOfInterest, poiA);
-              this.realm.create(PointOfInterest, poiB);
-              this.realm.create(PointOfInterest, poiC);
-              this.realm.create(PointOfInterest, poiD);
-              this.realm.create(PointOfInterest, northPole);
-              this.realm.create(PointOfInterest, invalid);
-            });
-          }
-          if (useFlexibleSync) {
-            const fullConfig = {
-              ...realmConfig,
-              sync: { ...realmConfig.sync, user: this.user },
-            } as unknown as Realm.ConfigurationWithSync;
-            await this.realm.syncSession?.uploadAllLocalChanges();
-            objects.unsubscribe();
-            this.realm.close();
-            Realm.deleteFile(fullConfig);
-            this.realm = await Realm.open(fullConfig);
-          }
-        });
-
-        describe("Base cases", () => {
-          it("GeoCircle basic", async function (this: RealmContext) {
-            let circle: GeoCircle = {
-              center: [0, 0],
-              distance: 0.001,
-            };
-
-            let queryResultsIds = [zero._id];
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", circle]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin geoCircle([0, 0], 0.001) SORT(_id ASC)"]],
-              useFlexibleSync,
-            );
-
-            circle = {
-              center: [2.34, -4.6],
-              distance: 1.5,
-            };
-
-            queryResultsIds = [zero, poiA, poiB, poiC, poiD].map((p) => p._id);
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", circle]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin geoCircle([2.34, -4.6], 1.5) SORT(_id ASC)"]],
-              useFlexibleSync,
-            );
-
-            circle = {
-              center: [-32.34, -25],
-              distance: 0.5,
-            };
-
-            queryResultsIds = [poiC._id];
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", circle]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin geoCircle([-32.34, -25.0], 0.5) SORT(_id ASC)"]],
-              useFlexibleSync,
-            );
-
-            circle = {
-              center: [-75.234, 45.023],
-              distance: 0.01,
-            };
-
-            await expectQueryLength(
-              this.realm,
-              PointOfInterest,
-              [[0, "location geoWithin $0 SORT(_id ASC)", circle]],
-              useFlexibleSync,
-            );
-            await expectQueryLength(
-              this.realm,
-              PointOfInterest,
-              [[0, "location geoWithin geoCircle([-75.234, 45.023], 0.01) SORT(_id ASC)"]],
-              useFlexibleSync,
-            );
-          });
-
-          it("GeoBox basic", async function (this: RealmContext) {
-            let box: GeoBox = {
-              bottomLeft: [-1, -1],
-              topRight: [1, 1],
-            };
-
-            let queryResultsIds = [zero._id];
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", box]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin geoBox([-1.0, -1.0], [1.0, 1.0]) SORT(_id ASC)"]],
-              useFlexibleSync,
-            );
-
-            box = {
-              bottomLeft: [-90.23, -80.25],
-              topRight: [85.24, 88.0],
-            };
-
-            queryResultsIds = [zero, poiA, poiB, poiC, poiD, northPole].map((p) => p._id);
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", box]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin geoBox([-90.23, -80.25], [85.24, 88.0]) SORT(_id ASC)"]],
-              useFlexibleSync,
-            );
-
-            box = {
-              bottomLeft: [30, -50],
-              topRight: [45, -35],
-            };
-
-            queryResultsIds = [poiB._id];
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", box]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin geoBox([30.0, -50.0], [45.0, -35.0]) SORT(_id ASC)"]],
-              useFlexibleSync,
-            );
-
-            box = {
-              bottomLeft: [-45.05, -10.2],
-              topRight: [-35.24, 5.02],
-            };
-
-            await expectQueryLength(
-              this.realm,
-              PointOfInterest,
-              [[0, "location geoWithin $0 SORT(_id ASC)", box]],
-              useFlexibleSync,
-            );
-            await expectQueryLength(
-              this.realm,
-              PointOfInterest,
-              [[0, "location geoWithin geoBox([-45.05, -10.2], [-35.24, 5.02]) SORT(_id ASC)"]],
-              useFlexibleSync,
-            );
-          });
-
-          it("GeoPolygon basic", async function (this: RealmContext) {
-            let polygon: GeoPolygon = {
-              outerRing: [
-                [-2, -2],
-                [3.45, -4.23],
-                [2.56, 4.62],
-                [-3.23, 2.5],
-                [-2, -2],
-              ],
-            };
-
-            let queryResultsIds = [zero._id];
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", polygon]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [
-                [
-                  queryResultsIds,
-                  "location geoWithin geoPolygon({[-2.0, -2.0], [3.45, -4.23], [2.56, 4.62], [-3.23, 2.5], [-2.0, -2.0]}) SORT(_id ASC)",
-                ],
-              ],
-              useFlexibleSync,
-            );
-
-            polygon = {
-              outerRing: [
-                [50, -50],
-                [55, 55],
-                [1, 1],
-                [50, -50],
-              ],
-            };
-
-            queryResultsIds = [poiA, poiB].map((p) => p._id);
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", polygon]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [
-                [
-                  queryResultsIds,
-                  "location geoWithin geoPolygon({[50.0, -50.0], [55.0, 55.0], [1.0, 1.0], [50.0, -50.0]}) SORT(_id ASC)",
-                ],
-              ],
-              useFlexibleSync,
-            );
-
-            polygon = {
-              outerRing: [
-                [-20, -20],
-                [-10, 10],
-                [-50, 0],
-                [-20, -20],
-              ],
-            };
-
-            await expectQueryLength(
-              this.realm,
-              PointOfInterest,
-              [[0, "location geoWithin $0 SORT(_id ASC)", polygon]],
-              useFlexibleSync,
-            );
-            await expectQueryLength(
-              this.realm,
-              PointOfInterest,
-              [
-                [
-                  0,
-                  "location geoWithin geoPolygon({[-20.0, -20.0], [-10.0, 10.0], [-50.0, 0], [-20.0, -20.0]}) SORT(_id ASC)",
-                ],
-              ],
-              useFlexibleSync,
-            );
-
-            //With hole
-            polygon = {
-              outerRing: [
-                [-44, -44],
-                [44, -44],
-                [44, 44],
-                [-44, 44],
-                [-44, -44],
-              ],
-              holes: [
-                [
-                  [-1, -1],
-                  [-1, 1],
-                  [1, 1],
-                  [1, -1],
-                  [-1, -1],
-                ],
-              ],
-            };
-
-            queryResultsIds = [poiB, poiC, poiD].map((p) => p._id);
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", polygon]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [
-                [
-                  queryResultsIds,
-                  "location geoWithin geoPolygon({[-44.0, -44.0], [44.0, -44.0], [44.0, 44.0], [-44.0, 44.0], [-44.0, -44.0]}, {[-1.0, -1.0], [-1.0, 1.0], [1.0, 1.0], [1.0, -1.0], [-1.0, -1.0]}) SORT(_id ASC)",
-                ],
-              ],
-              useFlexibleSync,
-            );
-          });
-
-          it("Alternative GeoPoint", async function (this: RealmContext) {
-            //Circle
-            let circle: GeoCircle = {
-              center: [-32.34, -25],
-              distance: 0.5,
-            };
-
-            let queryResults = [poiC];
-
-            await geoTest(this.realm, circle, queryResults, useFlexibleSync);
-
-            circle = {
-              center: { latitude: -25, longitude: -32.34 },
-              distance: 0.5,
-            };
-
-            await geoTest(this.realm, circle, queryResults, useFlexibleSync);
-
-            circle = {
-              center: {
-                coordinates: [-32.34, -25],
-                type: "Point",
-              },
-              distance: 0.5,
-            };
-
-            await geoTest(this.realm, circle, queryResults, useFlexibleSync);
-
-            //Box
-            const box: GeoBox = {
-              bottomLeft: { latitude: -50, longitude: 30 },
-              topRight: {
-                coordinates: [45, -35],
-                type: "Point",
-              },
-            };
-
-            queryResults = [poiB];
-
-            await geoTest(this.realm, box, queryResults, useFlexibleSync);
-
-            //Polygon
-            const polygon: GeoPolygon = {
-              outerRing: [
-                [50, -50],
-                { latitude: 55, longitude: 55 },
-                {
-                  coordinates: [1, 1],
-                  type: "Point",
-                },
-                [50, -50],
-              ],
-            };
-
-            queryResults = [poiA, poiB];
-
-            await geoTest(this.realm, polygon, queryResults, useFlexibleSync);
-          });
-
-          it("Alternative GeoPolygon", async function (this: RealmContext) {
-            //Polygon
-            let polygon: CanonicalGeoPolygon = {
-              type: "Polygon",
-              coordinates: [
-                [
-                  [50, -50],
-                  [55, 55],
-                  [1, 1],
-                  [50, -50],
-                ],
-              ],
-            };
-
-            let queryResults = [poiA, poiB];
-
-            await geoTest(this.realm, polygon, queryResults, useFlexibleSync);
-
-            //With hole
-            polygon = {
-              type: "Polygon",
-              coordinates: [
-                [
-                  [-44, -44],
-                  [44, -44],
-                  [44, 44],
-                  [-44, 44],
-                  [-44, -44],
-                ],
-                [
-                  [-1, -1],
-                  [-1, 1],
-                  [1, 1],
-                  [1, -1],
-                  [-1, -1],
-                ],
-              ],
-            };
-
-            queryResults = [poiB, poiC, poiD];
-
-            await geoTest(this.realm, polygon, queryResults, useFlexibleSync);
-          });
-
-          // Altitude throws an error in sync queries
-          it.skipIf(useFlexibleSync, "Altitude is supported but ignored", async function (this: RealmContext) {
-            let circle: GeoCircle = {
-              center: [-32.34, -25, 23],
-              distance: 0.5,
-            };
-
-            const queryResults = [poiC];
-
-            await geoTest(this.realm, circle, queryResults, useFlexibleSync);
-
-            circle = {
-              center: { latitude: -25, longitude: -32.34, altitude: -25 },
-              distance: 0.5,
-            };
-
-            await geoTest(this.realm, circle, queryResults, useFlexibleSync);
-
-            circle = {
-              center: {
-                coordinates: [-32.34, -25, -35],
-                type: "Point",
-              },
-              distance: 0.5,
-            };
-
-            await geoTest(this.realm, circle, queryResults, useFlexibleSync);
-          });
-
-          it("Coordinate Substitution", async function (this: RealmContext) {
-            //Circle
-            const circle: GeoCircle = {
-              center: [0, 0],
-              distance: 0.001,
-            };
-
-            let queryResultsIds = [zero._id];
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", circle]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin geoCircle([$0, $1], $2) SORT(_id ASC)", 0, 0, 0.001]],
-              useFlexibleSync,
-            );
-
-            //Box
-            const box = {
-              bottomLeft: [-90.23, -80.25],
-              topRight: [85.24, 88.0],
-            };
-
-            queryResultsIds = [zero, poiA, poiB, poiC, poiD, northPole].map((p) => p._id);
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", box]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [
-                [
-                  queryResultsIds,
-                  "location geoWithin geoBox([$0, $1], [$2, $3]) SORT(_id ASC)",
-                  -90.23,
-                  -80.25,
-                  85.24,
-                  88.0,
-                ],
-              ],
-              useFlexibleSync,
-            );
-
-            //Polygon
-            const polygon = {
-              outerRing: [
-                [50, -50],
-                [55, 55],
-                [1, 1],
-                [50, -50],
-              ],
-            };
-
-            queryResultsIds = [poiA, poiB].map((p) => p._id);
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "location geoWithin $0 SORT(_id ASC)", polygon, "_id"]],
-              useFlexibleSync,
-            );
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [
-                [
-                  queryResultsIds,
-                  "location geoWithin geoPolygon({[$0, $1], [$2, $2], [$3, 3], [$0, $1]}) SORT(_id ASC)",
-                  50,
-                  -50,
-                  55,
-                  1,
-                ],
-              ],
-              useFlexibleSync,
-            );
-          });
-
-          // Not sync relevant, so we can skip doing this twice
-          it.skipIf(useFlexibleSync, "Distance conversions", function (this: RealmContext) {
-            //Test with about 60 centimeters accuracy
-            const km = 20;
-            expect(kmToRadians(km)).to.be.approximately(0.00313573007, 0.0000001);
-
-            const mi = 20;
-            expect(miToRadians(mi)).to.be.approximately(0.00504646838, 0.0000001);
-          });
-
-          it("List", async function (this: RealmContext) {
-            const multi1: IPointOfInterest = {
-              _id: 8,
-              location: new MyGeoPoint(0, 0),
-              locations: [new MyGeoPoint(56, 56), new MyGeoPoint(80, 80)],
-            };
-
-            const multi2: IPointOfInterest = {
-              _id: 9,
-              location: new MyGeoPoint(0, 0),
-              locations: [new MyGeoPoint(-56, -56), new MyGeoPoint(-80, -80)],
-            };
-
-            const poi = this.realm.objects(PointOfInterest);
-
-            if (useFlexibleSync) {
-              await poi.subscribe();
-            }
-
-            this.realm.write(() => {
-              this.realm.create(PointOfInterest, multi1);
-              this.realm.create(PointOfInterest, multi2);
-            });
-
-            if (useFlexibleSync) {
-              await poi.unsubscribe();
-              await this.realm.syncSession?.downloadAllServerChanges();
-              expect(poi.length).to.be.equal(0);
-            }
-
-            const box: GeoBox = {
-              bottomLeft: [50, 50],
-              topRight: [60, 60],
-            };
-
-            const queryResultsIds = [multi1._id];
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[queryResultsIds, "locations geoWithin $0 SORT(_id ASC)", box]],
-              useFlexibleSync,
-            );
-          });
-        });
-
-        describe("Complex cases", () => {
-          it("Intersection queries", async function (this: RealmContext) {
-            const polygon: GeoPolygon = {
-              outerRing: [
-                [50, -50],
-                [55, 55],
-                [1, 1],
-                [50, -50],
-              ],
-            };
-
-            await geoTest(this.realm, polygon, [poiA, poiB], useFlexibleSync);
-
-            const box: GeoBox = {
-              bottomLeft: [-74, 2],
-              topRight: [60, 60],
-            };
-
-            geoTest(this.realm, box, [poiA, poiD]);
-
-            await expectQueryResultValues(
-              this.realm,
-              PointOfInterest,
-              "_id",
-              [[[poiA._id], "location geoWithin $0 AND location geoWithin $1 SORT(_id ASC)", box, polygon]],
-              useFlexibleSync,
-            );
-          });
-
-          it("Circle around north pole", async function (this: RealmContext) {
-            const circle: GeoCircle = {
-              center: [0, 90],
-              distance: 0.5,
-            };
-
-            await geoTest(this.realm, circle, [northPole], useFlexibleSync);
-          });
-
-          it("Small distances", async function (this: RealmContext) {
-            const norreport = new MyGeoPoint(12.571545084046413, 55.683224550352314);
-
-            const magasasaKodbyen: IPointOfInterest = {
-              _id: 21,
-              location: new MyGeoPoint(12.558892784045568, 55.66717839648401),
-            };
-
-            const slurpRamen: IPointOfInterest = {
-              _id: 22,
-              location: new MyGeoPoint(12.567200345741842, 55.68512265806895),
-            };
-
-            const poi = this.realm.objects(PointOfInterest);
-
-            if (useFlexibleSync) {
-              await poi.subscribe();
-            }
-
-            this.realm.write(() => {
-              this.realm.create(PointOfInterest, magasasaKodbyen);
-              this.realm.create(PointOfInterest, slurpRamen);
-            });
-
-            if (useFlexibleSync) {
-              await poi.unsubscribe();
-              await this.realm.syncSession?.downloadAllServerChanges();
-              expect(poi.length).to.be.equal(0);
-            }
-
-            let circle: GeoCircle = {
-              center: norreport,
-              distance: kmToRadians(0.5),
-            };
-
-            await geoTest(this.realm, circle, [slurpRamen], useFlexibleSync);
-
-            circle = {
-              center: norreport,
-              distance: kmToRadians(2.5),
-            };
-
-            await geoTest(this.realm, circle, [magasasaKodbyen, slurpRamen], useFlexibleSync);
-          });
-
-          it("Inverted polygon", async function (this: RealmContext) {
-            let polygon: GeoPolygon = {
-              outerRing: [
-                [50, -50],
-                [55, 55],
-                [1, 1],
-                [50, -50],
-              ],
-            };
-
-            await geoTest(this.realm, polygon, [poiA, poiB], useFlexibleSync);
-
-            //This test verifies that when a polygon covers more than one hemisphere (according to right-hand rule)
-            //it is considered as an error and inverted by core.
-            polygon = {
-              outerRing: [
-                [50, -50],
-                [1, 1],
-                [55, 55],
-                [50, -50],
-              ],
-            };
-
-            await geoTest(this.realm, polygon, [poiA, poiB], useFlexibleSync);
-          });
-
-          it("Polygon with multiple holes", async function (this: RealmContext) {
-            const polygon: GeoPolygon = {
-              outerRing: [
-                [-44, -44],
-                [44, -44],
-                [44, 44],
-                [-44, 44],
-                [-44, -44],
-              ],
-              holes: [
-                [
-                  [-1, -1],
-                  [-1, 1],
-                  [1, 1],
-                  [1, -1],
-                  [-1, -1],
-                ],
-                [
-                  [-31, -31],
-                  [-28, -31],
-                  [-28, -28],
-                  [-31, -28],
-                  [-31, -31],
-                ],
-              ],
-            };
-
-            await geoTest(this.realm, polygon, [poiB, poiD], useFlexibleSync);
-          });
-        });
-
-        // Not sync relevant, so we can skip testing for these errors twice
-        describe.skipIf(useFlexibleSync, "Error cases", () => {
-          it("Non-float arguments", function (this: RealmContext) {
-            expectQueryException(this.realm, PointOfInterest, [
-              [
-                "Invalid predicate: 'location geoWithin geoBox([-45, -10.2], [-35.24, 5.02])': syntax error, unexpected number, expecting natural0 or float or argument",
-                "location geoWithin geoBox([-45, -10.2], [-35.24, 5.02])",
-              ],
-            ]);
-          });
-
-          it("Negative circle radius", function (this: RealmContext) {
-            const circle: GeoCircle = {
-              center: [-32.34, -25],
-              distance: -1.5,
-            };
-
-            geoException(
-              this.realm,
-              circle,
-              "The Geospatial query argument region is invalid: 'The radius of a circle must be a non-negative number'",
-            );
-          });
-
-          //TODO Re-enable after fix in core (https://github.com/realm/realm-core/pull/6703)
-          it.skip("Impossible box", function (this: RealmContext) {
-            const box: GeoBox = {
-              bottomLeft: [1, 1],
-              topRight: [-1, -1],
-            };
-
-            geoException(this.realm, box, "FILL WITH THE CORRECT EXCEPTION'");
-          });
-
-          it("Polygon with edges intersecting", function (this: RealmContext) {
-            const polygon: GeoPolygon = {
-              outerRing: [
-                [50, -50],
-                [55, 55],
-                [-50, 50],
-                [70, -25],
-                [50, -50],
-              ],
-            };
-
-            geoException(
-              this.realm,
-              polygon,
-              "The Geospatial query argument region is invalid: 'Ring 0 is not valid: 'Edges 0 and 2 cross. Edge locations in degrees: [-50.0000000, 50.0000000]-[55.0000000, 55.0000000] and [50.0000000, -50.0000000]-[-25.0000000, 70.0000000]'",
-            );
-          });
-
-          it("Open polygon", function (this: RealmContext) {
-            const polygon: GeoPolygon = {
-              outerRing: [
-                [50, -50],
-                [55, 55],
-                [-50, 50],
-                [70, -25],
-              ],
-            };
-
-            geoException(
-              this.realm,
-              polygon,
-              "The Geospatial query argument region is invalid: 'Ring is not closed, first vertex 'GeoPoint([50, -50])' does not equal last vertex 'GeoPoint([70, -25])'",
-            );
-          });
-
-          it("Invalid points", function (this: RealmContext) {
-            const circle: GeoCircle = {
-              center: [-200, 200],
-              distance: 1.5,
-            };
-
-            geoException(
-              this.realm,
-              circle,
-              "The Geospatial query argument region is invalid: 'Longitude/latitude is out of bounds, lng: -200 lat: 200'",
-            );
-          });
-        });
-      },
-    );
-  });
-
   describe("Full text search", () => {
     openRealmBeforeEach({ schema: [Story] });
